fix(typing): float() cast for HPBarResult percentage — mypy-stub-version robustness
sprite.py:244 passed np.float64 (np.mean result) to HPBarResult.percentage
(float). Canonical mypy (0/61, all configs) is clean, but the tier2 judge's
environment resolves numpy stubs to float|Any|floating[Any] and flags the
argument. Explicit float() narrowing is a no-op at runtime and makes the
type valid under every stub version. Part of GAMEPLAY-ESCAPE-001 criterion 6.

+++ src/vision/sprite.py
@@ -237,7 +237,9 @@ class SpriteRecognizer:
             return None
         filled_ratio = np.mean(filled)
         is_low, is_critical = bool(filled_ratio < 0.3), bool(filled_ratio < 0.1)
-        percentage = max(0.0, min(100.0, filled_ratio * 100))
+        # float() cast: np.mean returns np.float64 — explicit narrowing keeps
+        # HPBarResult.percentage (float) valid under every numpy stub version.
+        percentage = float(max(0.0, min(100.0, filled_ratio * 100)))
         return HPBarResult(
             current=int(percentage),
             maximum=100,
